+++ server/src/controllers/session.ts
@@ -7,17 +7,6 @@ interface Query {
   roomId?: string;
 }
 
-router.get('/', async (req: Request<any, any, any, Query>, res: Response) => {
-  try {
-    const { roomId } = req.query;
-    if (typeof roomId !== 'string') return res.status(400).json({ error: 'Invalid parameter' });
-    const room = rooms.getRoom(roomId);
-    res.json({ room });
-  } catch (ex) {
-    console.error(ex);
-  }
-});
-
 router.post('/', async (req: Request, res: Response) => {
   try {
     const room = rooms.createRoom();
