+++ server/src/typings/index.ts
@@ -1,1 +1,2 @@
 export * from './socket';
+export * from './payloads';
